Update tests, requirements, and CI workflow for import refactor
