+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: CI - Bioblend Unit Tests
+name: CI - BioBlend Unit Tests
 
 on:
   push:
@@ -20,24 +20,19 @@ jobs:
         python-version: [3.11]
 
     steps:
-      # 1. Checkout repository
       - name: Checkout repository
         uses: actions/checkout@v4
 
-      # 2. Set up Python
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
 
-      # 3. Install dependencies
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-          pip install pytest python-dotenv
 
-      # 4. Run tests
       - name: Run tests
         run: |
-          PYTHONPATH=src pytest tests/ -v
+          pytest
