+++ requirements.txt
@@ -1,3 +1,3 @@
 bioblend>=1.7.0
 pytest>=9.0.0
-pytest-cov>=7.0.0
+
